move autocompleter into its own package
