+++ packages/autocomplete/.eslintignore
@@ -1,0 +1,3 @@
+node_modules
+local-test.js
+lib/
