+++ packages/autocomplete/AUTHORS
@@ -1,0 +1,4 @@
+Anna Herlihy <herlihyap@gmail.com>
+Durran Jordan <durran@gmail.com>
+Irina Shestak <shestak.irina@gmail.com>
+Maurizio Casimirri maurizio.cas@gmail.com
