+++ packages/autocomplete/index.spec.ts
@@ -1,23 +1,23 @@
-import completer from './completer';
+import autocomplete from './';
 import { signatures as shellSignatures } from '@mongosh/shell-api';
 
 import { expect } from 'chai';
 
-describe('completer.completer', () => {
+describe('autocomplete.autocomplete', () => {
   context('when context is top level shell api', () => {
     it('matches shell completions', () => {
       const i = 'u';
-      expect(completer('4.4.0', i)).to.deep.equal([['use'], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([['use'], i]);
     });
 
     it('does not have a match', () => {
       const i = 'ad';
-      expect(completer('4.4.0', i)).to.deep.equal([[], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([[], i]);
     });
 
     it('is an exact match to one of shell completions', () => {
       const i = 'use';
-      expect(completer('4.4.0', i)).to.deep.equal([[i], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([[i], i]);
     });
   });
 
@@ -27,19 +27,19 @@ describe('completer.completer', () => {
     // for now, this will only return the current input.
     it('matches a database command', () => {
       const i = 'db.agg';
-      expect(completer('4.4.0', i)).to.deep.equal([['db.aggregate'], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([['db.aggregate'], i]);
     });
 
     it('returns all suggestions', () => {
       const i = 'db.';
       const dbComplete = Object.keys(shellSignatures.Database.attributes);
       const adjusted = dbComplete.map(c => `${i}${c}`);
-      expect(completer('4.4.0', i)).to.deep.equal([adjusted, i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([adjusted, i]);
     });
 
     it('matches several suggestions', () => {
       const i = 'db.get';
-      expect(completer('4.4.0', i)[0]).to.include.members(
+      expect(autocomplete('4.4.0', i)[0]).to.include.members(
         [
           'db.getCollectionNames',
           'db.getCollection',
@@ -50,19 +50,19 @@ describe('completer.completer', () => {
 
     it('returns current input and no suggestions', () => {
       const i = 'db.shipw';
-      expect(completer('4.4.0', i)).to.deep.equal([[], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([[], i]);
     });
   });
 
   context('when context is collections', () => {
     it('matches a collection command', () => {
       const i = 'db.shipwrecks.findAnd';
-      expect(completer('4.4.0', i)).to.deep.equal([['db.shipwrecks.findAndModify'], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([['db.shipwrecks.findAndModify'], i]);
     });
 
     it('matches a collection command if part of an expression', () => {
       const i = 'var result = db.shipwrecks.findAnd';
-      expect(completer('4.4.0', i)).to.deep.equal([['var result = db.shipwrecks.findAndModify'], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([['var result = db.shipwrecks.findAndModify'], i]);
     });
 
     it('returns all suggestions', () => {
@@ -70,12 +70,12 @@ describe('completer.completer', () => {
       const collComplete = Object.keys(shellSignatures.Collection.attributes);
       const adjusted = collComplete.filter(c => !['count', 'update', 'save', 'remove'].includes(c)).map(c => `${i}${c}`);
 
-      expect(completer('4.4.0', i)).to.deep.equal([adjusted, i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([adjusted, i]);
     });
 
     it('matches several collection commands', () => {
       const i = 'db.shipwrecks.find';
-      expect(completer('4.4.0', i)).to.deep.equal([
+      expect(autocomplete('4.4.0', i)).to.deep.equal([
         [
           'db.shipwrecks.find', 'db.shipwrecks.findAndModify',
           'db.shipwrecks.findOne', 'db.shipwrecks.findOneAndDelete',
@@ -85,14 +85,14 @@ describe('completer.completer', () => {
 
     it('does not have a match', () => {
       const i = 'db.shipwrecks.pr';
-      expect(completer('4.4.0', i)).to.deep.equal([[], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([[], i]);
     });
   });
 
   context('when context is collections and aggregation cursor', () => {
     it('matches an aggregation cursor command', () => {
       const i = 'db.shipwrecks.aggregate([{$sort: {feature_type: 1}}]).has';
-      expect(completer('4.4.0', i)).to.deep.equal([
+      expect(autocomplete('4.4.0', i)).to.deep.equal([
         ['db.shipwrecks.aggregate([{$sort: {feature_type: 1}}]).hasNext'], i]);
     });
 
@@ -101,17 +101,17 @@ describe('completer.completer', () => {
       const aggCursorComplete = Object.keys(shellSignatures.AggregationCursor.attributes);
       const adjusted = aggCursorComplete.map(c => `${i}${c}`);
 
-      expect(completer('4.4.0', i)).to.deep.equal([adjusted, i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([adjusted, i]);
     });
 
     it('does not have a match', () => {
       const i = 'db.shipwrecks.aggregate([{$sort: {feature_type: 1}}]).w';
-      expect(completer('4.4.0', i)).to.deep.equal([[], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([[], i]);
     });
 
     it('has several matches', () => {
       const i = 'db.shipwrecks.aggregate([{$sort: {feature_type: 1}}]).i';
-      expect(completer('4.4.0', i)).to.deep.equal([
+      expect(autocomplete('4.4.0', i)).to.deep.equal([
         [
           'db.shipwrecks.aggregate([{$sort: {feature_type: 1}}]).isClosed',
           'db.shipwrecks.aggregate([{$sort: {feature_type: 1}}]).isExhausted',
@@ -123,19 +123,19 @@ describe('completer.completer', () => {
   context('when context is aggregation query', () => {
     it('has several matches', () => {
       const i = 'db.shipwrecks.aggregate([ { $so';
-      expect(completer('4.4.0', i)).to.deep.equal([
+      expect(autocomplete('4.4.0', i)).to.deep.equal([
         ['db.shipwrecks.aggregate([ { $sort',
           'db.shipwrecks.aggregate([ { $sortByCount'], i]);
     });
 
     it('does not have a match', () => {
       const i = 'db.shipwrecks.aggregate([ { $cat';
-      expect(completer('4.4.0', i)).to.deep.equal([[], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([[], i]);
     });
 
     it('matches an aggregation stage', () => {
       const i = 'db.shipwrecks.aggregate([ { $proj';
-      expect(completer('4.4.0', i)).to.deep.equal([
+      expect(autocomplete('4.4.0', i)).to.deep.equal([
         [ 'db.shipwrecks.aggregate([ { $project' ], i]);
     });
   });
@@ -143,7 +143,7 @@ describe('completer.completer', () => {
   context('when context is a collection query', () => {
     it('returns all suggestions', () => {
       const i = 'db.shipwrecks.find({ ';
-      expect(completer('4.4.0', i)[0]).to.include.members(
+      expect(autocomplete('4.4.0', i)[0]).to.include.members(
         [ 'db.shipwrecks.find({ $all',
           'db.shipwrecks.find({ $and',
           'db.shipwrecks.find({ $bitsAllClear',
@@ -193,7 +193,7 @@ describe('completer.completer', () => {
 
     it('has several matches', () => {
       const i = 'db.bios.find({ birth: { $g';
-      expect(completer('4.4.0', i)).to.deep.equal([
+      expect(autocomplete('4.4.0', i)).to.deep.equal([
         [
           'db.bios.find({ birth: { $geoIntersects',
           'db.bios.find({ birth: { $geoWithin',
@@ -204,12 +204,12 @@ describe('completer.completer', () => {
 
     it('does not have a match', () => {
       const i = 'db.bios.find({ field: { $cat';
-      expect(completer('4.4.0', i)).to.deep.equal([[], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([[], i]);
     });
 
     it('matches an aggregation stage', () => {
       const i = 'db.bios.find({ field: { $exis';
-      expect(completer('4.4.0', i)).to.deep.equal([
+      expect(autocomplete('4.4.0', i)).to.deep.equal([
         [ 'db.bios.find({ field: { $exists' ], i]);
     });
   });
@@ -217,7 +217,7 @@ describe('completer.completer', () => {
   context('when context is collections and collection cursor', () => {
     it('matches a collection cursor command', () => {
       const i = 'db.shipwrecks.find({feature_type: "Wrecks - Visible"}).for';
-      expect(completer('4.4.0', i)).to.deep.equal([
+      expect(autocomplete('4.4.0', i)).to.deep.equal([
         ['db.shipwrecks.find({feature_type: "Wrecks - Visible"}).forEach'], i]);
     });
 
@@ -257,7 +257,7 @@ describe('completer.completer', () => {
         'db.shipwrecks.find({feature_type: \"Wrecks - Visible\"}).toArray',
       ];
 
-      expect(completer('4.4.0', i)[0]).to.include.members(result);
+      expect(autocomplete('4.4.0', i)[0]).to.include.members(result);
     });
 
     it('returns all suggestions matching 3.0.0 version', () => {
@@ -293,17 +293,17 @@ describe('completer.completer', () => {
         'db.shipwrecks.find({feature_type: \"Wrecks - Visible\"}).toArray',
       ];
 
-      expect(completer('3.0.0', i)[0]).to.include.members(result);
+      expect(autocomplete('3.0.0', i)[0]).to.include.members(result);
     });
 
     it('does not have a match', () => {
       const i = 'db.shipwrecks.find({feature_type: "Wrecks - Visible"}).gre';
-      expect(completer('4.4.0', i)).to.deep.equal([[], i]);
+      expect(autocomplete('4.4.0', i)).to.deep.equal([[], i]);
     });
 
     it('has several matches', () => {
       const i = 'db.shipwrecks.find({feature_type: "Wrecks - Visible"}).cl';
-      expect(completer('4.4.0', i)).to.deep.equal([
+      expect(autocomplete('4.4.0', i)).to.deep.equal([
         [
           'db.shipwrecks.find({feature_type: "Wrecks - Visible"}).clone',
           'db.shipwrecks.find({feature_type: "Wrecks - Visible"}).close'
