+++ packages/autocomplete/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../../config/tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "./lib",
+    "allowJs": true
+  },
+  "include": [
+    "./*"
+  ],
+  "exclude": [
+    "./*.spec.*"
+  ]
+}
