+++ packages/browser-runtime-core/package.json
@@ -40,7 +40,7 @@
   "dependencies": {
     "@babel/generator": "^7.9.4",
     "@babel/parser": "^7.9.4",
-    "@mongosh/cli-repl": "^0.0.1-alpha.17",
+    "@mongosh/autocomplete": "^0.0.1-alpha.17",
     "@mongosh/service-provider-core": "^0.0.1-alpha.17",
     "@mongosh/shell-api": "^0.0.1-alpha.17",
     "@mongosh/shell-evaluator": "^0.0.1-alpha.17"
