+++ packages/cli-repl/package.json
@@ -38,6 +38,7 @@
     "@mongosh/i18n": "^0.0.1-alpha.17",
     "@mongosh/service-provider-server": "^0.0.1-alpha.17",
     "@mongosh/shell-api": "^0.0.1-alpha.17",
+    "@mongosh/autocomplete": "^0.0.1-alpha.17",
     "@mongosh/shell-evaluator": "^0.0.1-alpha.17",
     "acorn": "^7.1.1",
     "acorn-class-fields": "^0.3.2",
