+++ packages/cli-repl/src/index.ts
@@ -3,7 +3,6 @@ import redactPwd from './redact-pwd';
 import parseCliArgs from './arg-parser';
 import mapCliToDriver from './arg-mapper';
 import generateUri from './uri-generator';
-import completer from './completer';
 import clr from './clr';
 import { USAGE, TELEMETRY, MONGOSH_WIKI } from './constants';
 
@@ -16,7 +15,6 @@ export {
   MONGOSH_WIKI,
   CliRepl,
   redactPwd,
-  completer,
   generateUri,
   parseCliArgs,
   mapCliToDriver
